+++ src/machine_control/states_application/IdleState.cpp
@@ -46,7 +46,7 @@ bool IdleState::handleEvent(const EventPtr &event) {
 void IdleState::onReceivedConfigEvent(const EventPtr &event) {
   utils::Logger::log("-Handle event: kEventTypeReceivedConfig");
 
-  context.setConfigToSet(event->getArgumentAsType<uint32_t>());
+  context.setConfigToSet(event->getArgumentAsType<uint16_t>());
 
   auto state = std::make_shared<ConfigureState>(context);
   context.setCurrentState(state);
